Fixes TCP drops due to ARP timeouts
The root of the problem is that arpents were using u32s to store the
msec since boot.  Machines that run for over 50 days will wrap around.

Since NOW was a u64, the arpent ctimes were getting set to weird
numbers, and ultimately the arpents all appeared to be stale.

Stale arps get their blocklists pruned to one block (which is a
questionable bit of code).  These blocks were TCP packets.

So if a machine was running for more than 50 days, and if multiple TCP
packets were sent to the same arpent quickly, one of them could be
dropped, which throws off the entire TCP stream.

c89 happens to be a machine with a TSC that doesn't reset when
soft-rebooting (via our reboot()), so it acted like a machine that had
been up for about 3 months.

Arguably, we could have NOW (and the others, like seconds() and
milliseconds()) be cast to a u32, so that at least both NOW and ctime
are the same width, but that seems really shoddy.

+++ kern/include/ip.h
@@ -495,11 +495,11 @@ struct arpent {
 	struct arpent *hash;
 	struct block *hold;
 	struct block *last;
-	unsigned int ctime;			/* time entry was created or refreshed */
-	unsigned int utime;			/* time entry was last used */
+	uint64_t ctime;			/* time entry was created or refreshed */
+	uint64_t utime;			/* time entry was last used */
 	uint8_t state;
 	struct arpent *nextrxt;		/* re-transmit chain */
-	unsigned int rtime;			/* time for next retransmission */
+	uint64_t rtime;			/* time for next retransmission */
 	uint8_t rxtsrem;
 	struct Ipifc *ifc;
 	uint8_t ifcid;				/* must match ifc->id */
